+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 FEATURES:
 
 * **New Resource:** `agile_logical_port`
+* **New Resource:** `agile_logical_router`
 
 BUG FIXES:
 * Fix tenant always deploying without any change